Conflicts:
	molgenis-app/pom.xml
	molgenis-app/src/main/java/org/molgenis/app/WebAppConfig.java
	molgenis-catalogue/pom.xml
	molgenis-charts/pom.xml
	molgenis-core-ui/pom.xml
	molgenis-core/pom.xml
	molgenis-das/pom.xml
	molgenis-data-annotators/pom.xml
	molgenis-data-csv/pom.xml
	molgenis-data-elasticsearch/pom.xml
	molgenis-data-examples/pom.xml
	molgenis-data-excel/pom.xml
	molgenis-data-googlespreadsheet/pom.xml
	molgenis-data-import/pom.xml
	molgenis-data-jpa/pom.xml
	molgenis-data-mapper/pom.xml
	molgenis-data-merge/pom.xml
	molgenis-data-migrate/pom.xml
	molgenis-data-migrate/src/main/java/org/molgenis/migrate/version/MolgenisVersionService.java
	molgenis-data-mysql/pom.xml
	molgenis-data-rest-client/pom.xml
	molgenis-data-rest/pom.xml
	molgenis-data-rest/src/main/java/org/molgenis/data/rest/v2/RestControllerV2.java
	molgenis-data-semanticsearch/pom.xml
	molgenis-data-system/pom.xml
	molgenis-data-validation/pom.xml
	molgenis-data-vcf/pom.xml
	molgenis-data/pom.xml
	molgenis-data/src/main/java/org/molgenis/data/EntityMetaData.java
	molgenis-data/src/main/java/org/molgenis/data/meta/MetaDataServiceImpl.java
	molgenis-data/src/main/java/org/molgenis/data/support/AbstractEntityMetaData.java
	molgenis-dataexplorer/pom.xml
	molgenis-disease-matcher/pom.xml
	molgenis-file/pom.xml
	molgenis-js/pom.xml
	molgenis-model-registry/pom.xml
	molgenis-ontology-core/pom.xml
	molgenis-ontology/pom.xml
	molgenis-pathways/pom.xml
	molgenis-python/pom.xml
	molgenis-questionnaires/pom.xml
	molgenis-r/pom.xml
	molgenis-scripts-core/pom.xml
	molgenis-scripts/pom.xml
	molgenis-security-core/pom.xml
	molgenis-security/pom.xml
	pom.xml

+++ molgenis-data/src/main/java/org/molgenis/data/meta/MetaDataServiceImpl.java
@@ -1,16 +1,22 @@
 package org.molgenis.data.meta;
 
-import static com.google.common.collect.Lists.newArrayList;
 import static com.google.common.collect.Lists.reverse;
+import static org.molgenis.security.core.utils.SecurityUtils.getCurrentUsername;
+import static org.molgenis.util.SecurityDecoratorUtils.validatePermission;
 
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
+import java.util.function.Supplier;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
+import org.molgenis.MolgenisFieldTypes;
 import org.molgenis.data.AttributeMetaData;
 import org.molgenis.data.Entity;
 import org.molgenis.data.EntityMetaData;
@@ -22,13 +28,23 @@ import org.molgenis.data.RepositoryCollection;
 import org.molgenis.data.RepositoryDecorator;
 import org.molgenis.data.RepositoryDecoratorFactory;
 import org.molgenis.data.UnknownEntityException;
+import org.molgenis.data.i18n.I18nStringDecorator;
+import org.molgenis.data.i18n.I18nStringMetaData;
+import org.molgenis.data.i18n.LanguageMetaData;
+import org.molgenis.data.i18n.LanguageRepositoryDecorator;
+import org.molgenis.data.i18n.LanguageService;
+import org.molgenis.data.meta.system.ImportRunMetaData;
 import org.molgenis.data.support.DataServiceImpl;
 import org.molgenis.data.support.DefaultAttributeMetaData;
 import org.molgenis.data.support.DefaultEntityMetaData;
 import org.molgenis.data.support.NonDecoratingRepositoryDecoratorFactory;
+import org.molgenis.security.core.Permission;
+import org.molgenis.security.core.runas.RunAsSystem;
 import org.molgenis.security.core.runas.RunAsSystemProxy;
 import org.molgenis.security.core.utils.SecurityUtils;
 import org.molgenis.util.DependencyResolver;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.core.Ordered;
@@ -53,6 +69,8 @@ import com.google.common.collect.Sets;
  */
 public class MetaDataServiceImpl implements MetaDataService
 {
+	private static final Logger LOG = LoggerFactory.getLogger(MetaDataServiceImpl.class);
+
 	private PackageRepository packageRepository;
 	private EntityMetaDataRepository entityMetaDataRepository;
 	private AttributeMetaDataRepository attributeMetaDataRepository;
@@ -60,10 +78,17 @@ public class MetaDataServiceImpl implements MetaDataService
 	private final Map<String, RepositoryCollection> backends = Maps.newHashMap();
 	private final DataServiceImpl dataService;
 	private TransactionTemplate transactionTemplate;
+	private LanguageService languageService;
 
 	public MetaDataServiceImpl(DataServiceImpl dataService)
 	{
 		this.dataService = dataService;
+	}
+
+	@Autowired
+	public void setLanguageService(LanguageService languageService)
+	{
+		this.languageService = languageService;
 	}
 
 	@Autowired
@@ -86,10 +111,14 @@ public class MetaDataServiceImpl implements MetaDataService
 		this.defaultBackend = backend;
 		backends.put(backend.getName(), backend);
 
+		I18nStringMetaData.INSTANCE.setBackend(backend.getName());
+		LanguageMetaData.INSTANCE.setBackend(backend.getName());
 		PackageMetaData.INSTANCE.setBackend(backend.getName());
 		TagMetaData.INSTANCE.setBackend(backend.getName());
 		EntityMetaDataMetaData.INSTANCE.setBackend(backend.getName());
 		AttributeMetaDataMetaData.INSTANCE.setBackend(backend.getName());
+
+		ImportRunMetaData.INSTANCE.setBackend(backend.getName());
 
 		bootstrapMetaRepos();
 		return this;
@@ -97,21 +126,49 @@ public class MetaDataServiceImpl implements MetaDataService
 
 	private void bootstrapMetaRepos()
 	{
+		Repository languageRepo = defaultBackend.addEntityMeta(LanguageMetaData.INSTANCE);
+		dataService.addRepository(new LanguageRepositoryDecorator(languageRepo, dataService));
+
+		Repository i18StringsRepo = defaultBackend.addEntityMeta(I18nStringMetaData.INSTANCE);
+		dataService.addRepository(new I18nStringDecorator(i18StringsRepo));
+
+		Supplier<Stream<String>> languageCodes = () -> languageService.getLanguageCodes().stream();
+
+		// Add language attributes to the AttributeMetaDataMetaData
+		languageCodes.get().map(code -> AttributeMetaDataMetaData.LABEL + '-' + code)
+				.forEach(AttributeMetaDataMetaData.INSTANCE::addAttribute);
+
+		// Add description attributes to the AttributeMetaDataMetaData
+		languageCodes.get().map(code -> AttributeMetaDataMetaData.DESCRIPTION + '-' + code)
+				.forEach(attrName -> AttributeMetaDataMetaData.INSTANCE.addAttribute(attrName)
+						.setDataType(MolgenisFieldTypes.TEXT));
+
+		// Add description attributes to the EntityMetaDataMetaData
+		languageCodes.get().map(code -> EntityMetaDataMetaData.DESCRIPTION + '-' + code)
+				.forEach(attrName -> EntityMetaDataMetaData.INSTANCE.addAttribute(attrName)
+						.setDataType(MolgenisFieldTypes.TEXT));
+
+		// Add language attributes to the EntityMetaDataMetaData
+		languageCodes.get().map(code -> EntityMetaDataMetaData.LABEL + '-' + code)
+				.forEach(EntityMetaDataMetaData.INSTANCE::addAttribute);
+
+		// Add language attributes to I18nStringMetaData
+		languageCodes.get().forEach(I18nStringMetaData.INSTANCE::addLanguage);
 
 		Repository tagRepo = defaultBackend.addEntityMeta(TagMetaData.INSTANCE);
 		dataService.addRepository(tagRepo);
 
 		Repository packages = defaultBackend.addEntityMeta(PackageRepository.META_DATA);
-		dataService.addRepository(packages);
+		dataService.addRepository(new MetaDataRepositoryDecorator(packages));
 		packageRepository = new PackageRepository(packages);
 
-		attributeMetaDataRepository = new AttributeMetaDataRepository(defaultBackend);
+		attributeMetaDataRepository = new AttributeMetaDataRepository(defaultBackend, languageService);
 		entityMetaDataRepository = new EntityMetaDataRepository(defaultBackend, packageRepository,
-				attributeMetaDataRepository);
+				attributeMetaDataRepository, languageService);
 		attributeMetaDataRepository.setEntityMetaDataRepository(entityMetaDataRepository);
 
-		dataService.addRepository(attributeMetaDataRepository.getRepository());
-		dataService.addRepository(entityMetaDataRepository.getRepository());
+		dataService.addRepository(new MetaDataRepositoryDecorator(attributeMetaDataRepository.getRepository()));
+		dataService.addRepository(new MetaDataRepositoryDecorator(entityMetaDataRepository.getRepository()));
 		entityMetaDataRepository.fillEntityMetaDataCache();
 	}
 
@@ -133,6 +190,8 @@ public class MetaDataServiceImpl implements MetaDataService
 	@Override
 	public void deleteEntityMeta(String entityName)
 	{
+		validatePermission(entityName, Permission.WRITEMETA);
+
 		transactionTemplate.execute((TransactionStatus status) -> {
 			EntityMetaData emd = getEntityMetaData(entityName);
 			if ((emd != null) && !emd.isAbstract())
@@ -147,6 +206,7 @@ public class MetaDataServiceImpl implements MetaDataService
 		});
 
 		refreshCaches();
+		LOG.info("Repository [{}] deleted by user [{}]", entityName, getCurrentUsername());
 	}
 
 	private void deleteEntityPermissions(String entityName)
@@ -156,14 +216,14 @@ public class MetaDataServiceImpl implements MetaDataService
 		// User permissions
 		if (dataService.hasRepository("UserAuthority"))
 		{
-			Iterable<Entity> userPermissions = dataService.query("UserAuthority").in("role", authorities).findAll();
+			Stream<Entity> userPermissions = dataService.query("UserAuthority").in("role", authorities).findAll();
 			dataService.delete("UserAuthority", userPermissions);
 		}
 
 		// Group permissions
 		if (dataService.hasRepository("GroupAuthority"))
 		{
-			Iterable<Entity> groupPermissions = dataService.query("GroupAuthority").in("role", authorities).findAll();
+			Stream<Entity> groupPermissions = dataService.query("GroupAuthority").in("role", authorities).findAll();
 			dataService.delete("GroupAuthority", groupPermissions);
 		}
 	}
@@ -172,6 +232,8 @@ public class MetaDataServiceImpl implements MetaDataService
 	@Override
 	public void delete(List<EntityMetaData> entities)
 	{
+		entities.forEach(emd -> validatePermission(emd.getName(), Permission.WRITEMETA));
+
 		reverse(DependencyResolver.resolve(Sets.newHashSet(entities))).stream().map(EntityMetaData::getName)
 				.forEach(this::deleteEntityMeta);
 	}
@@ -183,6 +245,8 @@ public class MetaDataServiceImpl implements MetaDataService
 	@Override
 	public void deleteAttribute(String entityName, String attributeName)
 	{
+		validatePermission(entityName, Permission.WRITEMETA);
+
 		// Update AttributeMetaDataRepository
 		entityMetaDataRepository.removeAttribute(entityName, attributeName);
 		EntityMetaData emd = getEntityMetaData(entityName);
@@ -192,8 +256,8 @@ public class MetaDataServiceImpl implements MetaDataService
 	private ManageableRepositoryCollection getManageableRepositoryCollection(EntityMetaData emd)
 	{
 		RepositoryCollection backend = getBackend(emd);
-		if (!(backend instanceof ManageableRepositoryCollection)) throw new RuntimeException(
-				"Backend  is not a ManageableCrudRepositoryCollection");
+		if (!(backend instanceof ManageableRepositoryCollection))
+			throw new RuntimeException("Backend  is not a ManageableCrudRepositoryCollection");
 
 		return (ManageableRepositoryCollection) backend;
 	}
@@ -210,9 +274,8 @@ public class MetaDataServiceImpl implements MetaDataService
 
 	@Transactional
 	@Override
-	public Repository add(EntityMetaData emd, RepositoryDecoratorFactory decoratorFactory)
+	public synchronized Repository add(EntityMetaData emd, RepositoryDecoratorFactory decoratorFactory)
 	{
-
 		MetaValidationUtils.validateEntityMetaData(emd);
 		RepositoryCollection backend = getBackend(emd);
 
@@ -223,9 +286,8 @@ public class MetaDataServiceImpl implements MetaDataService
 			if (!dataService.hasRepository(emd.getName()))
 			{
 				Repository repo = backend.getRepository(emd.getName());
-				if (repo == null) throw new UnknownEntityException(String.format(
-						"Unknown entity '%s' for backend '%s'", emd.getName(), backend.getName()));
-
+				if (repo == null) throw new UnknownEntityException(
+						String.format("Unknown entity '%s' for backend '%s'", emd.getName(), backend.getName()));
 				Repository decoratedRepo = decoratorFactory.createDecoratedRepository(repo);
 
 				dataService.addRepository(decoratedRepo);
@@ -250,6 +312,7 @@ public class MetaDataServiceImpl implements MetaDataService
 
 		Repository repo = backend.addEntityMeta(getEntityMetaData(emd.getName()));
 		Repository decoratedRepo = decoratorFactory.createDecoratedRepository(repo);
+
 		dataService.addRepository(decoratedRepo);
 
 		// Return decorated repo
@@ -258,32 +321,16 @@ public class MetaDataServiceImpl implements MetaDataService
 
 	@Transactional
 	@Override
-	public synchronized Repository addEntityMeta(EntityMetaData emd)
+	public Repository addEntityMeta(EntityMetaData emd)
 	{
-		if (emd.getDecorator() == null)
-		{
-			return add(emd, new NonDecoratingRepositoryDecoratorFactory());
-		}
-		else
-		{
-			return add(emd, new RepositoryDecoratorFactory()
-			{
-				@Override
-				public Repository createDecoratedRepository(Repository repository)
-				{
-					RepositoryDecorator repoDecorator = emd.getDecorator();
-					repoDecorator.setTarget(repository);
-
-					return repoDecorator;
-				}
-			});
-		}
+		return add(emd, new NonDecoratingRepositoryDecoratorFactory());
 	}
 
 	@Transactional
 	@Override
 	public void addAttribute(String fullyQualifiedEntityName, AttributeMetaData attr)
 	{
+		validatePermission(fullyQualifiedEntityName, Permission.WRITEMETA);
 		MetaValidationUtils.validateName(attr.getName());
 
 		EntityMetaData emd = entityMetaDataRepository.addAttribute(fullyQualifiedEntityName, attr);
@@ -293,6 +340,7 @@ public class MetaDataServiceImpl implements MetaDataService
 	@Override
 	public void addAttributeSync(String fullyQualifiedEntityName, AttributeMetaData attr)
 	{
+		validatePermission(fullyQualifiedEntityName, Permission.WRITEMETA);
 		MetaValidationUtils.validateName(attr.getName());
 
 		EntityMetaData emd = entityMetaDataRepository.addAttribute(fullyQualifiedEntityName, attr);
@@ -341,7 +389,10 @@ public class MetaDataServiceImpl implements MetaDataService
 	@Transactional
 	public void recreateMetaDataRepositories()
 	{
-		delete(newArrayList(getEntityMetaDatas()));
+		List<EntityMetaData> metas = getDefaultBackend().getRepository(EntityMetaDataMetaData.ENTITY_NAME).stream()
+				.map(e -> getEntityMetaData(e.getString(EntityMetaDataMetaData.FULL_NAME)))
+				.collect(Collectors.toList());
+		delete(metas);
 
 		attributeMetaDataRepository.deleteAll();
 		entityMetaDataRepository.deleteAll();
@@ -393,12 +444,15 @@ public class MetaDataServiceImpl implements MetaDataService
 	}
 
 	@Override
+	@RunAsSystem
 	public synchronized void onApplicationEvent(ContextRefreshedEvent event)
 	{
 		// Discover all backends
-		Map<String, RepositoryCollection> backendBeans = event.getApplicationContext().getBeansOfType(
-				RepositoryCollection.class);
+		Map<String, RepositoryCollection> backendBeans = event.getApplicationContext()
+				.getBeansOfType(RepositoryCollection.class);
 		backendBeans.values().forEach(this::addBackend);
+
+		Map<String, EntityMetaData> emds = event.getApplicationContext().getBeansOfType(EntityMetaData.class);
 
 		// Create repositories from EntityMetaData in EntityMetaData repo
 		for (EntityMetaData emd : entityMetaDataRepository.getMetaDatas())
@@ -413,16 +467,23 @@ public class MetaDataServiceImpl implements MetaDataService
 		}
 
 		// Discover static EntityMetaData
-		Map<String, EntityMetaData> emds = event.getApplicationContext().getBeansOfType(EntityMetaData.class);
-		DependencyResolver.resolve(Sets.newHashSet(emds.values())).forEach(this::addEntityMeta);
-		// Update update manageable backends, JPA throws exception
-		DependencyResolver.resolve(Sets.newHashSet(emds.values())).stream().filter(this::isManageableBackend)
-				.forEach(this::updateEntityMeta);
+		Set<EntityMetaData> staticEmd = new HashSet<>(emds.values());
+		Set<EntityMetaData> allEmd = new HashSet<>(getEntityMetaDatas());
+		allEmd.addAll(staticEmd);
+
+		// Use all EntityMetaData for dependency resolving
+		List<EntityMetaData> resolved = DependencyResolver.resolve(allEmd);
+		resolved.retainAll(staticEmd);// Only keep the static EntityMetaData
+
+		resolved.stream().filter(emd -> !dataService.hasRepository(emd.getName())).forEach(this::addEntityMeta);
+
+		// Update update manageable backends
+		resolved.stream().filter(this::isManageableBackend).forEach(this::updateEntityMeta);
 	}
 
 	private boolean isManageableBackend(EntityMetaData emd)
 	{
-		// Might work for more than just MySQL backend, but not JPA backend
+		// Might work for more than just MySQL backend
 		return emd.getBackend() == null || "MySql".equals(emd.getBackend());
 	}
 
@@ -434,6 +495,8 @@ public class MetaDataServiceImpl implements MetaDataService
 
 	public void updateEntityMetaBackend(String entityName, String backend)
 	{
+		validatePermission(entityName, Permission.WRITEMETA);
+
 		DefaultEntityMetaData entityMeta = entityMetaDataRepository.get(entityName);
 		if (entityMeta == null) throw new UnknownEntityException("Unknown entity '" + entityName + "'");
 		entityMeta.setBackend(backend);
@@ -450,10 +513,9 @@ public class MetaDataServiceImpl implements MetaDataService
 	public LinkedHashMap<String, Boolean> integrationTestMetaData(RepositoryCollection repositoryCollection)
 	{
 		LinkedHashMap<String, Boolean> entitiesImportable = new LinkedHashMap<String, Boolean>();
-		StreamSupport.stream(repositoryCollection.getEntityNames().spliterator(), false).forEach(
-				entityName -> entitiesImportable.put(entityName, this
-						.canIntegrateEntityMetadataCheck(repositoryCollection.getRepository(entityName)
-								.getEntityMetaData())));
+		StreamSupport.stream(repositoryCollection.getEntityNames().spliterator(), false)
+				.forEach(entityName -> entitiesImportable.put(entityName, this.canIntegrateEntityMetadataCheck(
+						repositoryCollection.getRepository(entityName).getEntityMetaData())));
 
 		return entitiesImportable;
 	}
@@ -465,11 +527,9 @@ public class MetaDataServiceImpl implements MetaDataService
 	{
 		LinkedHashMap<String, Boolean> entitiesImportable = new LinkedHashMap<String, Boolean>();
 
-		StreamSupport.stream(newEntitiesMetaDataMap.keySet().spliterator(), false).forEach(
-				entityName -> entitiesImportable.put(
-						entityName,
-						skipEntities.contains(entityName)
-								|| this.canIntegrateEntityMetadataCheck(newEntitiesMetaDataMap.get(entityName))));
+		StreamSupport.stream(newEntitiesMetaDataMap.keySet().spliterator(), false)
+				.forEach(entityName -> entitiesImportable.put(entityName, skipEntities.contains(entityName)
+						|| this.canIntegrateEntityMetadataCheck(newEntitiesMetaDataMap.get(entityName))));
 
 		return entitiesImportable;
 	}
@@ -482,13 +542,13 @@ public class MetaDataServiceImpl implements MetaDataService
 			EntityMetaData newEntity = newEntityMetaData;
 			EntityMetaData oldEntity = dataService.getEntityMetaData(entityName);
 
-			List<AttributeMetaData> oldAtomicAttributes = StreamSupport.stream(
-					oldEntity.getAtomicAttributes().spliterator(), false).collect(
-					Collectors.<AttributeMetaData> toList());
+			List<AttributeMetaData> oldAtomicAttributes = StreamSupport
+					.stream(oldEntity.getAtomicAttributes().spliterator(), false)
+					.collect(Collectors.<AttributeMetaData> toList());
 
 			LinkedHashMap<String, AttributeMetaData> newAtomicAttributesMap = new LinkedHashMap<String, AttributeMetaData>();
-			StreamSupport.stream(newEntity.getAtomicAttributes().spliterator(), false).forEach(
-					attribute -> newAtomicAttributesMap.put(attribute.getName(), attribute));
+			StreamSupport.stream(newEntity.getAtomicAttributes().spliterator(), false)
+					.forEach(attribute -> newAtomicAttributesMap.put(attribute.getName(), attribute));
 
 			for (AttributeMetaData oldAttribute : oldAtomicAttributes)
 			{
@@ -503,5 +563,11 @@ public class MetaDataServiceImpl implements MetaDataService
 		}
 
 		return true;
+	}
+
+	@Override
+	public boolean hasBackend(String backendName)
+	{
+		return backends.containsKey(backendName);
 	}
 }
